Removed calculate_scores (converted to function) from BaseSimilarity
